implement the changes in hp and covid jsons
