+++ packages/esm-commons-lib/src/utils/schema-manipulation.ts
@@ -1,0 +1,48 @@
+export function extractDefaults(schema) {
+  const result = {};
+
+  function traverse(schema) {
+    for (const key in schema) {
+      if (key === '_default') {
+        Object.assign(result, schema[key]);
+      } else if (typeof schema[key] === 'object' && !Array.isArray(schema[key])) {
+        traverse(schema[key]);
+      }
+    }
+  }
+
+  traverse(schema);
+  return result;
+}
+
+export function replaceWithConfigDefaults(obj, configDefaults) {
+  if (Array.isArray(obj)) {
+    return obj.map((item) => {
+      if (typeof item === 'string' && configDefaults.hasOwnProperty(item)) {
+        return configDefaults[item];
+      } else {
+        return replaceWithConfigDefaults(item, configDefaults);
+      }
+    });
+  } else if (typeof obj === 'object' && obj !== null) {
+    const newObj = {};
+    for (const key in obj) {
+      if (typeof obj[key] === 'string') {
+        if (configDefaults.hasOwnProperty(key)) {
+          // Case where UUID is the value
+          newObj[configDefaults[key]] = obj[key];
+        } else if (configDefaults.hasOwnProperty(obj[key])) {
+          // Case where UUID is the key
+          newObj[key] = configDefaults[obj[key]];
+        } else {
+          newObj[key] = obj[key];
+        }
+      } else {
+        newObj[key] = replaceWithConfigDefaults(obj[key], configDefaults);
+      }
+    }
+    return newObj;
+  } else {
+    return obj;
+  }
+}
