+++ packages/esm-covid-app/src/config-schema.ts
@@ -21,10 +21,7 @@ export const configSchema = {
       covidTypeofTestConcept_UUID: '069f6dfe-88c1-4a45-a894-0d99549c8718',
       covidSpecimenCollectionDate_UUID: '159951AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
       covidOutcomeUUID: 'a845f3e6-4432-4de4-9fff-37fa270b1a06',
-      covidTestType: '069f6dfe-88c1-4a45-a894-0d99549c8718',
-      covidOutcome: 'a845f3e6-4432-4de4-9fff-37fa270b1a06',
       rapidAntigenResultDate: 'af159c77-bc5d-46dd-90d9-bcbffb22267f',
-      covidTreatementOutConcept_UUID: 'a845f3e6-4432-4de4-9fff-37fa270b1a06',
       covidRapidTestResultDate_UUID: 'af159c77-bc5d-46dd-90d9-bcbffb22267f',
       pcrTestResultDate: '4a77ab44-0323-490e-96be-e168c0e5c9de',
       finalCovid19Result: '5da5c21b-969f-41bd-9091-e40d4c707544',
@@ -32,9 +29,7 @@ export const configSchema = {
       covidDiagnorticPcrResultDate_UUID: '4a77ab44-0323-490e-96be-e168c0e5c9de',
       htsRetrospectiveEncounterType: '79c1f50f-f77d-42e2-ad2a-d29304dde2fe',
       htsRetrospectiveType: '79c1f50f-f77d-42e2-ad2a-d29304dde2fe',
-      pcrTestResult: '3f4ee14b-b4ab-4597-9fe9-406883b63d76',
       covidReasonsForTestingConcep_UUID: 'ae46f4b1-c15d-4bba-ab41-b9157b82b0ce',
-      covidDiagnosticPcrResult_UUID: '3f4ee14b-b4ab-4597-9fe9-406883b63d76',
       covidTestResultUUID: '3f4ee14b-b4ab-4597-9fe9-406883b63d76',
       covidPatientStatusUUID: 'de3bc9b7-05b5-41b6-a38d-8d2eec646c4f',
       covidSARS_TestResultConcept_UUID: '89feed9c-1dd9-477a-ab1c-86f5f75f6762',
@@ -66,6 +61,10 @@ export const configSchema = {
       covidVaccineAdministeredConcept_UUID: 'e41fbe17-4aee-4a44-950b-6676d6e0ede2',
       covidVaccineConcept_UUID: '0cc868bd-e9dd-4b59-b278-f923afe22d82',
       covidVaccineSeriesConcept_UUID: '45b3959d-f897-4e01-b6e5-16ef6ae0a687',
+      covidNotDoneStatus: '1118AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+      covidCompletedStatus: '1267AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+      covidCancelledStatus: '165170AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
+      covidPendingStatus: '162866AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
     },
   },
   cohorts: {
@@ -109,7 +108,6 @@ export const configSchema = {
       covidLabResultFormUuid: 'cf8cd756-baef-38df-b84d-92968b42c113',
       covidLabCancellationFormUuid: 'd38bc949-c95b-39eb-a2c5-08b82a36409c',
       covidSampleCollectionFormUuid: '371d19b6-485f-11ec-99cc-1fdd2d4e9d88',
-      covidLabOrderFormUuid: 'f5fb6bc4-6fc3-3462-a191-2fff0896bab3',
     },
   },
 };
