+++ packages/esm-covid-app/src/views/patient chart/covid-vaccinations.component.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { EncounterList, getMenuItemTabConfiguration } from '@ohri/openmrs-esm-ohri-commons-lib';
 
 import covidVaccinationsSchemaConfig from './covid-vaccinations-schema.json';
+import { configSchema } from '../../config-schema';
 
 interface CovidVaccinationsWidgetProps {
   patientUuid: string;
@@ -10,7 +11,7 @@ interface CovidVaccinationsWidgetProps {
 export const covidFormSlot = 'hts-encounter-form-slot';
 
 const CovidVaccinations: React.FC<CovidVaccinationsWidgetProps> = ({ patientUuid }) => {
-  const tabs = getMenuItemTabConfiguration(covidVaccinationsSchemaConfig);
+  const tabs = getMenuItemTabConfiguration(covidVaccinationsSchemaConfig, configSchema);
 
   return (
     <>
