+++ packages/esm-hiv-prevention-app/src/views/hiv-testing-services/hts-prevention-schema-config.json
@@ -5,13 +5,13 @@
       "tabName": "HIV Testing",
       "headerTitle": "HIV Testing",
       "displayText": "HIV Testing",
-      "encounterType": "79c1f50f-f77d-42e2-ad2a-d29304dde2fe",
+      "encounterType": "htsRetrospectiveEncounterType",
       "columns": [
         {
           "id": "date",
           "isDate": true,
           "title": "Date of HIV Test",
-          "concept": "164400AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA"
+          "concept": "dateOfHIVTestingConceptUUID"
         },
         {
           "id": "location",
@@ -21,7 +21,7 @@
         {
           "id": "hivTestResult",
           "title": "HIV Test result",
-          "concept": "e16b0068-b6a2-46b7-aba9-e3be00a7b4ab"
+          "concept": "hivTestResultConceptUUID"
         },
         {
           "id": "provider",
@@ -33,13 +33,13 @@
           "title": "Actions",
           "actionOptions": [
             {
-              "formName": "HIV Testing",
+              "formName": "HIVTestingFormName",
               "package": "hiv-prevention",
               "label": "View Details",
               "mode": "view"
             },
             {
-              "formName": "HIV Testing",
+              "formName": "HIVTestingFormName",
               "package": "hiv-prevention",
               "label": "Edit Form",
               "mode": "edit"
@@ -53,8 +53,8 @@
       },
       "formList": [
         {
-          "name": "HIV Testing",
-          "uuid": "43ffea77-49dc-3ebd-8d83-c2aedb654030",
+          "name": "HIVTestingFormName",
+          "uuid": "hivTestingFormUuid",
           "excludedIntents": ["HTS_PRE_TEST", "HTS_TEST", "HTS_POST_TEST"],
           "fixedIntent": "*"
         }
